changed all int to double

+++ src/conditioning.cc
@@ -160,7 +160,7 @@ void Conditioning::BeginProcessing_() {
   }
 }
 
-typedef std::map<std::string, std::map<std::string, int>> package_;
+typedef std::map<std::string, std::map<std::string, double>> package_;
 
 //- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 void Conditioning::PackageMatl_(int pack_size, package_ package_prop) {
